Rename 'Mock' classes to 'Fake' classes, plus package name
'Mocks' usually include expectation tests for how they are called; these
classes simply provide sparse-but-functioning implementations of
objects normally provided by the HSM.
